Safari-611.1.10/LayoutTests/accessibility/ios-simulator/blockquote-level.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>

<blockquote><div role="text" id="item1">1</div></blockquote>

<blockquote><blockquote><div role="text" id="item2">2</div></blockquote></blockquote>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that blockquote level is correctly exposed to the iOS accessibility API.");
    
    if (window.accessibilityController) {
        var item1 = accessibilityController.accessibleElementById("item1");
        shouldBe("item1.numberAttributeValue('AXBlockquoteLevel')", "1");

        var item2 = accessibilityController.accessibleElementById("item2");
        shouldBe("item2.numberAttributeValue('AXBlockquoteLevel')", "2");
    }
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
