Safari-611.1.10/LayoutTests/accessibility/ios-simulator/blockquote-level-expected.txt

1
2
This tests that blockquote level is correctly exposed to the iOS accessibility API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item1.numberAttributeValue('AXBlockquoteLevel') is 1
PASS item2.numberAttributeValue('AXBlockquoteLevel') is 2
PASS successfullyParsed is true

TEST COMPLETE
